0.2.6.1018     - (berrinam) [Internal] : Remove Commandline field from Job     - (berrinam) [Internal] : Remove commandline generation from CommandLineGenerator and put it in individual encoders     - (berrinam) [Internal] : Remove Framecount and FPS from VideoJob     - (berrinam) [Internal] : Separated Job into Job Queue info (now under TaggedJob) and job running info (under Job, as before). This may cause temporary Serialization bugs     - (berrinam) [Internal] : Reworked/simplified BitrateCalculator     - (berrinam) [Main window, Mux window, OneClick] : Can have arbitrarily many audio tracks     - (berrinam) [Internal] : Documentation for MuxPath finding     - (berrinam) [Internal] : Moved drag & drop code to FileBar to avoid repetition     - (berrinam) Load jobs now works for mux and audio jobs
git-svn-id: http://svn.code.sf.net/p/megui/code@230 a21b8521-4df1-4ed3-b48c-cff5bc73dcb3

+++ trunk/packages/tools/autoencode/AutoEncodeWindow.cs
@@ -544,7 +544,7 @@ namespace MeGUI
                     getAudioStreamsForBitrate(),
                     desiredSizeBytes,
                     videoStream.NumberOfFrames,
-                    videoStream.Framerate,
+                    (double)videoStream.Framerate,
                     out videoSize);
 #warning check whether codecs use k=1000 or k=1024 for kbits // kick those that still use 1024...
                 this.videoSize.Text = new FileSize(Unit.KB, videoSize).ToString();
@@ -581,7 +581,7 @@ namespace MeGUI
                     getAudioStreamsForBitrate(),
                     desiredBitrate,
                     videoStream.NumberOfFrames,
-                    videoStream.Framerate,
+                    (double)videoStream.Framerate,
                     out rawVideoSize) / 1024L ;
                 this.videoSize.Text = new FileSize(Unit.KB, rawVideoSize).ToString();
                 this.targetSize.Value = new FileSize(Unit.MB, outputSize);
@@ -824,7 +824,7 @@ namespace MeGUI
                 myVideo.Input = info.Video.Info.VideoInput;
                 myVideo.Output = info.Video.Info.VideoOutput;
                 myVideo.NumberOfFrames = length;
-                myVideo.Framerate = framerate;
+                myVideo.Framerate = (decimal)framerate;
                 myVideo.DAR = info.Video.Info.DAR;
                 myVideo.VideoType = info.Video.CurrentMuxableVideoType;
                 myVideo.Settings = vSettings;
